refactor: make LSP client separate from transport (#83)
This PR factors out the shared interface and implementation
but provides two constructors for TCP and WS based LSP
clients
