+++ pkg/flink/lsp/lsp_completer_test.go
@@ -20,7 +20,7 @@ func TestLSPIntialize(t *testing.T) {
 	uri := lsp.DocumentURI("file:///test.sql")
 
 	lspClient := &LSPClient{documentURI: &uri, conn: conn}
-	_, err := lspClient.initialize()
+	_, err := lspClient.Initialize()
 	require.NoError(t, err)
 }
 
@@ -30,7 +30,7 @@ func TestLSPIntializeCallErr(t *testing.T) {
 	uri := lsp.DocumentURI("file:///test.sql")
 
 	lspClient := &LSPClient{documentURI: &uri, conn: conn}
-	_, err := lspClient.initialize()
+	_, err := lspClient.Initialize()
 	require.Error(t, err)
 }
 
@@ -38,7 +38,7 @@ func TestLSPIntializeNoConnErr(t *testing.T) {
 	uri := lsp.DocumentURI("file:///test.sql")
 
 	lspClient := &LSPClient{documentURI: &uri, conn: nil}
-	_, err := lspClient.initialize()
+	_, err := lspClient.Initialize()
 	require.Error(t, err)
 }
 
@@ -48,7 +48,7 @@ func TestLSPdidOpen(t *testing.T) {
 
 	lspClient := &LSPClient{conn: conn}
 	require.Nil(t, lspClient.documentURI)
-	err := lspClient.didOpen()
+	err := lspClient.DidOpen()
 	require.NotNil(t, lspClient.documentURI)
 	require.NoError(t, err)
 }
@@ -59,40 +59,40 @@ func TestLSPdidOpenCallErr(t *testing.T) {
 
 	lspClient := &LSPClient{conn: conn}
 	require.Nil(t, lspClient.documentURI)
-	err := lspClient.didOpen()
+	err := lspClient.DidOpen()
 	require.Nil(t, lspClient.documentURI)
 	require.Error(t, err)
 }
 
 func TestLSPdidOpenNoConnErr(t *testing.T) {
 	lspClient := &LSPClient{conn: nil}
-	err := lspClient.didOpen()
+	err := lspClient.DidOpen()
 	require.Error(t, err)
 }
 
 func TestLSPdidChange(t *testing.T) {
 	uri := lsp.DocumentURI("file:///test.sql")
 	conn := mock.NewMockJSONRpcConn(gomock.NewController(t))
-	conn.EXPECT().Call(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Return(nil).Times(1)
+	conn.EXPECT().Notify(gomock.Any(), "textDocument/didChange", gomock.Any()).Return(nil).Times(1)
 
 	lspClient := &LSPClient{documentURI: &uri, conn: conn}
-	err := lspClient.didChange("some text")
+	err := lspClient.DidChange("some text")
 	require.NoError(t, err)
 }
 
 func TestLSPdidChangeCallErr(t *testing.T) {
 	uri := lsp.DocumentURI("file:///test.sql")
 	conn := mock.NewMockJSONRpcConn(gomock.NewController(t))
-	conn.EXPECT().Call(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).Return(errors.New("some error")).Times(1)
+	conn.EXPECT().Notify(gomock.Any(), "textDocument/didChange", gomock.Any()).Return(errors.New("some error")).Times(1)
 
 	lspClient := &LSPClient{documentURI: &uri, conn: conn}
-	err := lspClient.didChange("some text")
+	err := lspClient.DidChange("some text")
 	require.Error(t, err)
 }
 
 func TestLSPdidChangeNoConnErr(t *testing.T) {
 	lspClient := &LSPClient{conn: nil}
-	err := lspClient.didChange("some text")
+	err := lspClient.DidChange("some text")
 	require.Error(t, err)
 }
 
@@ -103,10 +103,10 @@ func TestLSPCompletion(t *testing.T) {
 	s := store.NewStore(mock.NewFakeFlinkGatewayClient(), func() {}, &types.ApplicationOptions{}, func() error { return nil })
 
 	lspClient := &LSPClient{documentURI: &uri, conn: conn, store: s}
-	completion, err := lspClient.completion(lsp.Position{})
+	Completion, err := lspClient.Completion(lsp.Position{})
 
 	require.NoError(t, err)
-	require.NotNil(t, completion)
+	require.NotNil(t, Completion)
 }
 
 func TestLSPCompletionCallErr(t *testing.T) {
@@ -116,22 +116,22 @@ func TestLSPCompletionCallErr(t *testing.T) {
 	s := store.NewStore(mock.NewFakeFlinkGatewayClient(), func() {}, &types.ApplicationOptions{}, func() error { return nil })
 
 	lspClient := &LSPClient{documentURI: &uri, conn: conn, store: s}
-	completion, err := lspClient.completion(lsp.Position{})
+	Completion, err := lspClient.Completion(lsp.Position{})
 
 	require.Error(t, err)
-	require.Equal(t, lsp.CompletionList{}, completion)
+	require.Equal(t, lsp.CompletionList{}, Completion)
 }
 
 func TestLSPCompletionNoConnErr(t *testing.T) {
 	uri := lsp.DocumentURI("file:///test.sql")
 
 	lspClient := &LSPClient{documentURI: &uri, conn: nil}
-	_, err := lspClient.completion(lsp.Position{})
+	_, err := lspClient.Completion(lsp.Position{})
 	require.Error(t, err)
 }
 
 func TestNewLSPClient(t *testing.T) {
-	lspClient := NewLSPClient(nil).(*LSPClient)
+	lspClient := NewLocalLSPClient(nil).(*LSPClient)
 	require.NotNil(t, lspClient)
 
 	isRunning := waitForConditionWithTimeout(func() bool { return lspClient.conn != nil }, 3*time.Second)
